+++ src/usettings.pas
@@ -13,7 +13,6 @@ type
   { TFSettings }
 
   TFSettings = class(TForm)
-    BBOSMCache: TBitBtn;
     BPDefaultButtons: TButtonPanel;
     GroupBox1: TGroupBox;
     GroupBox2: TGroupBox;
@@ -30,7 +29,9 @@ type
     LELongitude: TLabeledEdit;
     LEMapCache: TLabeledEdit;
     SDDCacheDirectory: TSelectDirectoryDialog;
+    SpeedButton1: TSpeedButton;
     procedure BBOSMCacheClick(Sender: TObject);
+    procedure BPDefaultButtonsClick(Sender: TObject);
     procedure CancelButtonClick(Sender: TObject);
     procedure OKButtonClick(Sender: TObject);
   private
@@ -54,6 +55,12 @@ begin
   if SDDCacheDirectory.Execute then
     LEMapCache.Caption := SDDCacheDirectory.FileName;
 end;
+
+procedure TFSettings.BPDefaultButtonsClick(Sender: TObject);
+begin
+
+end;
+
 
 procedure TFSettings.CancelButtonClick(Sender: TObject);
 begin