+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## v0.3.0
 
 - ADD: Support for Status reports.
 - FIX: Create PoI from APRS data received over FlexPacket.